Initial Working Release
This is the initial working release. However, it won’t be published in
its current state. The readme.txt file needs updating and old commented
out code needs to be removed. It also needs to be moved into a plugin
directory within the Flat Bootstrap theme and activated by that theme.